+++ endpoints/instantiations/auth_/signup.py
@@ -1,0 +1,3 @@
+from utilities.namespace import create_namespace, AppNamespaces
+
+namespace_signup = create_namespace(AppNamespaces.AUTH)
